Code/requirements.txt
tensorflow>=2.10.0
opencv-python>=4.5.0
pillow>=9.0.0
numpy>=1.21.0
scipy>=1.7.0
